microblog/db_upgrade.py
#!flask/bin/python
from migrate.versioning import api
from config import SQLALCHEMY_DATABASE_URI
from config import SQLALCHEMY_MIGRATE_REPO
api.upgrade(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
#print 'Current database version: ' + str(api.db_version(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO))
# If you have database migration support, then when you are ready to release 
# the new version of the app to your production server you just need to record 
# a new migration, copy the migration scripts to your production server and run 
# a simple script that applies the changes for you. 